+++ packages/types/src/interfaces/accounts/definitions.ts
@@ -1,4 +1,4 @@
 import { Definitions } from '@polkadot/types/types';
-import accounts from '@reef-defi/type-definitions/accounts';
+import accounts from '@dust-defi/type-definitions/accounts';
 
 export default accounts as Definitions;