+++ packages/types/src/interfaces/support/definitions.ts
@@ -1,4 +1,4 @@
 import { Definitions } from '@polkadot/types/types';
-import support from '@reef-defi/type-definitions/support';
+import support from '@dust-defi/type-definitions/support';
 
 export default support as Definitions;