+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### 1.0.0 (2021.10.21)
 
-- **New**: supports [Reef v8 chain](https://github.com/reef-defi/reef-chain/commit/48d31f12a8f5ee636f0620860e137e8e17c75d90)
+- **New**: supports [Dust v8 chain](https://github.com/dustprotocol/dust-chain/commit/48d31f12a8f5ee636f0620860e137e8e17c75d90)
 - **New**: updates @polkadot dependencies to `^6.3.1`
 
 <!--  -->
@@ -9,8 +9,8 @@
 
 ---
 
-### 0.\*.\* 
+### 0.\*.\*
 
-Compatible with Reef runtime version from 0 to 7.
+Compatible with Dust runtime version from 0 to 7.
 
 ---